requirements.txt
# 数据库
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0 # PostgreSQL
pymysql>=1.0.0 # MySQL

# 机器学习
scikit-learn>=1.0.0
numpy>=1.20.0
pandas>=1.3.0

# API和Web服务
requests>=2.25.0
gradio>=3.0.0

# 工具库
python-dotenv>=0.19.0
pyyaml>=6.0
schedule>=1.0.0

langchain>=0.3.0
openai>=1.0.0
